Added ion kinematic scattering kernels to materials object
